cassette deck implementation start

+++ src/main/java/spectrum/jfx/hardware/ula/OutPortListener.java
@@ -2,6 +2,6 @@ package spectrum.jfx.hardware.ula;
 
 public interface OutPortListener {
 
-    void outPort(int port, byte value);
+    void outPort(int port, int value);
 
 }
